Restructure package and update tests
