+++ src/monster_rpg/items/equipment.py
