+++ src/monster_rpg/items/item_data.py
