+++ src/monster_rpg/map/locations.json
