+++ src/monster_rpg/map_data.py
