+++ src/monster_rpg/monsters/synthesis_rules.py
