+++ src/monster_rpg/start_rpg.py
@@ -1,6 +1,6 @@
 """Convenience script to launch the Monster RPG web interface."""
 
-from web_main import app
+from .web_main import app
 
 if __name__ == "__main__":
     app.run(debug=True)
