+++ src/monster_rpg/static/styles.css
