+++ src/monster_rpg/templates/battle.html
