+++ src/monster_rpg/templates/battle_log.html
