+++ src/monster_rpg/templates/explore.html
