+++ src/monster_rpg/templates/formation.html
