+++ src/monster_rpg/templates/index.html
