+++ src/monster_rpg/templates/items.html
