+++ src/monster_rpg/templates/layout.html
