+++ src/monster_rpg/templates/map.html
