+++ src/monster_rpg/templates/monster_book.html
