+++ src/monster_rpg/templates/party.html
