+++ src/monster_rpg/templates/play.html
