+++ src/monster_rpg/templates/result.html
