+++ src/monster_rpg/templates/shop.html
