+++ src/monster_rpg/templates/status.html
