+++ src/monster_rpg/templates/synthesize.html
