+++ src/monster_rpg/web_main.py
@@ -17,13 +17,13 @@ except ImportError as e:  # pragma: no cover - dependency check
     ) from e
 import random
 
-import database_setup
 import sqlite3
-from player import Player
-from monsters.monster_data import ALL_MONSTERS, MONSTER_BOOK_DATA
-from items.item_data import ALL_ITEMS
-from map_data import LOCATIONS, get_map_overview, get_map_grid, load_locations
-from exploration import generate_enemy_party
+from . import database_setup
+from .player import Player
+from .monsters.monster_data import ALL_MONSTERS, MONSTER_BOOK_DATA
+from .items.item_data import ALL_ITEMS
+from .map_data import LOCATIONS, get_map_overview, get_map_grid, load_locations
+from .exploration import generate_enemy_party
 
 app = Flask(__name__)
 app.secret_key = "dev-secret"
