+++ tests/test_enemy_weights.py
@@ -1,11 +1,7 @@
-import os
-import sys
 import random
 import unittest
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-
-from map_data import load_locations, LOCATIONS
+from monster_rpg.map_data import load_locations, LOCATIONS
 
 class EnemyWeightSelectionTests(unittest.TestCase):
     def test_weighted_selection_deterministic(self):
