+++ tests/test_item_monster_synthesis.py
@@ -1,11 +1,7 @@
-import os
-import sys
 import unittest
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-
-from player import Player
-from items.item_data import ALL_ITEMS
+from monster_rpg.player import Player
+from monster_rpg.items.item_data import ALL_ITEMS
 
 class ItemMonsterSynthesisTests(unittest.TestCase):
     def test_monster_item_fusion(self):
