+++ tests/test_level_up.py
@@ -1,13 +1,8 @@
-import os
-import sys
 import unittest
 import io
 from contextlib import redirect_stdout
 
-# Ensure repo root is on path
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-
-from monsters.monster_data import SLIME
+from monster_rpg.monsters.monster_data import SLIME
 
 class LevelUpSpeedTests(unittest.TestCase):
     def test_speed_increase_and_message(self):
