+++ tests/test_party_buff.py
@@ -1,12 +1,8 @@
-import os
-import sys
 import unittest
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-
-from battle import apply_skill_effect, process_status_effects
-from monsters.monster_class import Monster
-from skills.skills import ALL_SKILLS
+from monster_rpg.battle import apply_skill_effect, process_status_effects
+from monster_rpg.monsters.monster_class import Monster
+from monster_rpg.skills.skills import ALL_SKILLS
 
 class PartyBuffTests(unittest.TestCase):
     def test_brave_song_buff_applies_to_all_allies(self):
